+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPQP8K7CZG29DFD8WJ1MR3VZ"
+        "01KQ6F63YVKTQDJGGW4JPC3CF0"
     )
     r"""The unique identifier of the data source"""