+++ packages/orq-rc/src/orq_ai_sdk/routingrules.py
@@ -311,7 +311,7 @@ class RoutingRules(BaseSDK):
             return unmarshal_json_response(
                 models.RoutingRuleCreateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "409", "4XX"], "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -429,7 +429,7 @@ class RoutingRules(BaseSDK):
             return unmarshal_json_response(
                 models.RoutingRuleCreateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "409", "4XX"], "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -901,7 +901,7 @@ class RoutingRules(BaseSDK):
             return unmarshal_json_response(
                 models.RoutingRuleUpdateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "409", "4XX"], "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -1025,7 +1025,7 @@ class RoutingRules(BaseSDK):
             return unmarshal_json_response(
                 models.RoutingRuleUpdateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "409", "4XX"], "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):